Updated user /feed delete command

+++ cogs/feeds/cog.py
@@ -94,14 +94,19 @@ class Feeds(commands.Cog, name="Feeds"):
             }
         em = bake_embed(embed)
         
-        try:
-            logging.warning(f"Deleting entry {feed_id} from task database!")
-            db_tasks.open.delete_one({"internal_id": feed_id})
-            await interaction.send(embed=em, ephemeral=True)
-        except Exception as e:
-            logging.exception(e)
+        # TODO: Add check, if feed is actually from the user!
+        if db_tasks.open.find_one({"user_id": interaction.user.id, "internal_id": feed_id}):
+            try:
+                logging.warning(f"Deleting entry {feed_id} from task database!")
+                db_tasks.open.delete_one({"internal_id": feed_id})
+                await interaction.send(embed=em, ephemeral=True)
+            except Exception as e:
+                logging.exception(e)
+                await interaction.send(embed=em_error(), ephemeral=True)
+                return
+        else:
+            # TODO: Replace with actual error message
             await interaction.send(embed=em_error(), ephemeral=True)
-            return
     
     
     # Command subgroup: /feed admin ...
